refactor(ui): split integration tauri services
