+++ src/ui/src/services/tauri/remote.ts
@@ -1,1 +1,49 @@
-export {};
+import { invoke } from "@tauri-apps/api/core";
+import type {
+  DiscoveredServer,
+  PairResult,
+  RemoteConnectionInfo,
+  RemoteInitialData,
+} from "../../types/remote";
+
+export function listRemoteConnections(): Promise<RemoteConnectionInfo[]> {
+  return invoke("list_remote_connections");
+}
+
+export function pairWithServer(
+  host: string,
+  port: number,
+  pairingToken: string
+): Promise<PairResult> {
+  return invoke("pair_with_server", { host, port, pairingToken });
+}
+
+export function connectRemote(id: string): Promise<RemoteInitialData | null> {
+  return invoke("connect_remote", { id });
+}
+
+export function disconnectRemote(id: string): Promise<void> {
+  return invoke("disconnect_remote", { id });
+}
+
+export function removeRemoteConnection(id: string): Promise<void> {
+  return invoke("remove_remote_connection", { id });
+}
+
+export function listDiscoveredServers(): Promise<DiscoveredServer[]> {
+  return invoke("list_discovered_servers");
+}
+
+export function addRemoteConnection(
+  connectionString: string
+): Promise<PairResult> {
+  return invoke("add_remote_connection", { connectionString });
+}
+
+export function sendRemoteCommand(
+  connectionId: string,
+  method: string,
+  params: Record<string, unknown>
+): Promise<unknown> {
+  return invoke("send_remote_command", { connectionId, method, params });
+}
