+++ src/ui/src/services/tauri/scm.ts
@@ -1,1 +1,35 @@
-export {};
+import { invoke } from "@tauri-apps/api/core";
+import type { PluginInfo, PullRequest, ScmDetail } from "../../types/plugin";
+
+export function listScmProviders(): Promise<PluginInfo[]> {
+  return invoke("list_scm_providers");
+}
+
+export function getScmProvider(repoId: string): Promise<string | null> {
+  return invoke("get_scm_provider", { repoId });
+}
+
+export function setScmProvider(repoId: string, pluginName: string): Promise<void> {
+  return invoke("set_scm_provider", { repoId, pluginName });
+}
+
+export function loadScmDetail(workspaceId: string): Promise<ScmDetail> {
+  return invoke("load_scm_detail", { workspaceId });
+}
+
+export function scmCreatePr(
+  workspaceId: string,
+  title: string,
+  body: string,
+  base: string,
+  draft: boolean
+): Promise<PullRequest> {
+  return invoke("scm_create_pr", { workspaceId, title, body, base, draft });
+}
+
+export function scmMergePr(
+  workspaceId: string,
+  prNumber: number
+): Promise<unknown> {
+  return invoke("scm_merge_pr", { workspaceId, prNumber });
+}
